iabstract svg - broke inertia drag bounds
