+++ README.md
@@ -6,3 +6,7 @@ A React version of Chris Gannon's [SVG Bubble Slider](https://codepen.io/chrisga
 
 [react-svg-bubble-slider.netlify.app](https://react-svg-bubble-slider.netlify.app/?path=/docs/intro--page)
 `
+
+## Licence
+
+GSAP and all GreenSock files are subject to the standard GreenSock license which can be found at [https://greensock.com/standard-license](https://greensock.com/standard-license)
