+++ package.json
@@ -34,7 +34,6 @@
   "contributors": [
     "Chris Gannon"
   ],
-  "license": "MIT",
   "bugs": {
     "url": "https://github.com/PaulieScanlon/react-svg-bubble-slider/issues"
   },
