+++ src/components/SvgBubbleSlider/SvgBubbleSlider.stories.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import { SvgBubbleSlider } from '.'
 
+import { Button } from 'theme-ui'
+
 export default {
   title: 'Props',
   parameters: {
@@ -10,6 +12,20 @@ export default {
 }
 
 export const usage = () => <SvgBubbleSlider />
+
+export const action = () => (
+  <SvgBubbleSlider>
+    {({ reaction }) => <Button>{`Post:${reaction}`}</Button>}
+  </SvgBubbleSlider>
+)
+
+action.story = {
+  parameters: {
+    docs: {
+      storyDescription: 'The `scale` prop can be used to set the size',
+    },
+  },
+}
 
 export const scale = () => <SvgBubbleSlider scale="50%" />
